started adding functionality around position awareness
Started playing with the 'PY' command to get the current position in
steps of the motor. Right now, the GUI will display the current position
of the motor (as determined by getting the current step value and
converting it to mm) when the motor is activated, and after hitting the
stop button. I need to figure out a way to get the program to constantly
update the position without interrupting the communication between the
GUI and the controller...